refactor(onboard): split person assembly into _onboard_person.ts and fix review issues
- Extract person-mode assembly to _onboard_person.ts (file was 707 lines, >500 limit)
- Add module-level constants for projection kinds instead of inline SQL strings
- Fix getContradictionProjections to use buildStaleMap for real stale values
- Escape % and _ in user-supplied paths for LIKE queries (injection safety)
- Use RELATION_TYPES.LIKELY_OWNER_OF instead of inline literal
